Fix dependency requirement and add more documentation
Because this plugin just supports the interface defined by `TypedStruct`, we don't actually have to
require a specific version of the calling library.

Which is good because the original library seems to have become stale and the author has not responded
to requests to add maintainers.

This PR removes that dependency requirement, and adds documentation about installation and basic usage